// src/payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "employees".
 */
export interface Employee {
  id: number;
  name: string;
  email?: string | null;
  department?: string | null;
  designation?: string | null;
  joinedAt?: string | null;
  photo?: (number | null) | Media;
  updatedAt: string;
  createdAt: string;
}